+++ walking-skeleton/server/repositories/postRepository.js
@@ -2,8 +2,8 @@ import postgres from "postgres";
 
 const sql = postgres();
 
-const create = async (communityId, post) => {
-    const result = await sql`INSERT INTO posts (community_id, title, content) VALUES (${communityId}, ${post.title}, ${post.content}) RETURNING *;`;
+const create = async (userId, communityId, post) => {
+    const result = await sql`INSERT INTO posts (created_by, community_id, title, content) VALUES (${userId}, ${communityId}, ${post.title}, ${post.content}) RETURNING *;`;
     return result[0];
 };
 
@@ -16,8 +16,8 @@ const findById = async (communityId, postId) => {
     return result[0];
 };
 
-const deleteById = async (communityId, postId) => {
-    const result = await sql`DELETE FROM posts WHERE community_id = ${communityId} AND id = ${postId} RETURNING *;`;
+const deleteById = async (userId, communityId, postId) => {
+    const result = await sql`DELETE FROM posts WHERE created_by = ${userId} AND community_id = ${communityId} AND id = ${postId} RETURNING *;`;
     return result[0];
 };
